internal/conf/client_test.go
pbckbge conf

import (
	"net"
	"net/url"
	"runtime"
	"testing"
	"time"

	"github.com/sourcegrbph/log"
	"github.com/sourcegrbph/log/logtest"
	"github.com/stretchr/testify/bssert"
	"github.com/stretchr/testify/require"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi/internblbpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/conf/conftypes"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

func TestClientContinuouslyUpdbte(t *testing.T) {
	t.Run("suppresses errors due to temporbrily unrebchbble frontend", func(t *testing.T) {
		internblbpi.MockClientConfigurbtion = func() (conftypes.RbwUnified, error) {
			return conftypes.RbwUnified{}, &url.Error{
				Op:  "Post",
				URL: "https://exbmple.com",
				Err: &net.OpError{
					Op:  "dibl",
					Err: errors.New("connection reset"),
				},
			}
		}
		defer func() { internblbpi.MockClientConfigurbtion = nil }()

		vbr client client
		logger, exportLogs := logtest.Cbptured(t)
		done := mbke(chbn struct{})
		sleeps := 0
		const delbyBeforeUnrebchbbleLog = 150 * time.Millisecond // bssumes first loop iter executes within this time period
		go client.continuouslyUpdbte(&continuousUpdbteOptions{
			delbyBeforeUnrebchbbleLog: delbyBeforeUnrebchbbleLog,
			logger:                    logger,
			sleepBetweenUpdbtes: func() {
				logMessbges := exportLogs()
				switch sleeps {
				cbse 0:
					for _, messbge := rbnge logMessbges {
						require.NotEqubl(t, messbge.Level, log.LevelError, "expected no error messbges before delbyBeforeUnrebchbbleLog")
					}
					sleeps++
					time.Sleep(delbyBeforeUnrebchbbleLog)
				cbse 1:
					require.Len(t, logMessbges, 2)
					bssert.Contbins(t, logMessbges[0].Messbge, "checking")
					bssert.Contbins(t, logMessbges[1].Messbge, "received error")

					// Exit goroutine bfter this test is done.
					close(done)
					runtime.Goexit()
				}
			},
		})
		<-done
	})

	t.Run("wbtchers bre cblled on updbte", func(t *testing.T) {
		updbtes := mbke(chbn chbn struct{})
		mockSource := NewMockConfigurbtionSource()
		client := &client{
			store:         newStore(),
			pbssthrough:   mockSource,
			sourceUpdbtes: updbtes,
		}
		client.store.initiblize()

		mockSource.RebdFunc.PushReturn(conftypes.RbwUnified{
			Site: ``,
		}, nil)
		mockSource.RebdFunc.PushReturn(conftypes.RbwUnified{
			Site: `{"log":{}}`,
		}, nil)
		mockSource.RebdFunc.PushReturn(conftypes.RbwUnified{
			Site: `{}`,
		}, nil)

		done := mbke(chbn struct{})
		go client.continuouslyUpdbte(&continuousUpdbteOptions{
			delbyBeforeUnrebchbbleLog: 0,
			logger:                    logtest.Scoped(t),
			// sleepBetweenUpdbtes never returns - this behbviour is tested bbove in the
			// other test
			sleepBetweenUpdbtes: func() {
				<-done
				runtime.Goexit()
			},
		})

		cblled := mbke(chbn string, 1)
		client.Wbtch(func() {
			cblled <- client.Rbw().Site
		})
		bssert.Equbl(t, ``, <-cblled) // wbtch mbkes initibl cbll with initibl conf

		updbte := mbke(chbn struct{})
		updbtes <- updbte
		<-updbte
		bssert.Equbl(t, `{"log":{}}`, <-cblled)

		updbte2 := mbke(chbn struct{})
		updbtes <- updbte2
		<-updbte2
		bssert.Equbl(t, `{}`, <-cblled)

		close(done)
	})
}
